--- ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/sync/SavePreMergeHeadersStep.java
@@ -1,124 +1,0 @@
-/*
- * Copyright contributors to Hyperledger Besu.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-package org.hyperledger.besu.ethereum.eth.sync;
-
-import static org.hyperledger.besu.util.log.LogUtil.throttledLog;
-
-import org.hyperledger.besu.ethereum.ConsensusContext;
-import org.hyperledger.besu.ethereum.chain.MutableBlockchain;
-import org.hyperledger.besu.ethereum.core.BlockHeader;
-import org.hyperledger.besu.ethereum.core.Difficulty;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.Function;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-/** A step in the synchronization process that saves historical block headers. */
-public class SavePreMergeHeadersStep
-    implements Function<List<BlockHeader>, CompletableFuture<List<BlockHeader>>> {
-  private static final Logger LOG = LoggerFactory.getLogger(SavePreMergeHeadersStep.class);
-  private final MutableBlockchain blockchain;
-  private final boolean isPoS;
-  private final long lastPoWBlockNumber;
-  private final Optional<ConsensusContext> consensusContext;
-
-  private final AtomicBoolean shouldLog = new AtomicBoolean(true);
-  private static final int LOG_REPEAT_DELAY_SECONDS = 30;
-
-  public SavePreMergeHeadersStep(
-      final MutableBlockchain blockchain,
-      final boolean isPoS,
-      final long checkpointBlockNumber,
-      final Optional<ConsensusContext> consensusContext) {
-    this.blockchain = blockchain;
-    this.isPoS = isPoS;
-    this.lastPoWBlockNumber = checkpointBlockNumber - 1;
-    this.consensusContext = consensusContext;
-  }
-
-  @Override
-  public CompletableFuture<List<BlockHeader>> apply(final List<BlockHeader> blockHeaders) {
-    if (blockHeaders.getLast().getNumber() > lastPoWBlockNumber) {
-      // we are not pos or all headers are post-merge, so we can just pass them through
-      return CompletableFuture.completedFuture(blockHeaders);
-    } else if (blockHeaders.getFirst().getNumber() < lastPoWBlockNumber) {
-      // all headers are post-merge, so we can just pass them through
-      return storeBlockHeaders(blockHeaders);
-    } else {
-      // the last PoW header is included, so we store all headers up to and including it and we pass
-      // the rest through
-      final List<BlockHeader> allHeaders = blockHeaders.stream().toList();
-      final int preMergeEnd = (int) (lastPoWBlockNumber - allHeaders.getFirst().getNumber() + 1);
-      final List<BlockHeader> preMergeHeaders = allHeaders.subList(0, preMergeEnd);
-      final List<BlockHeader> postMergeHeaders = allHeaders.subList(preMergeEnd, allHeaders.size());
-      return storeBlockHeaders(preMergeHeaders)
-          .thenRun(
-              () -> {
-                if (isPoS) {
-                  consensusContext.ifPresent(
-                      context ->
-                          blockchain
-                              .getTotalDifficultyByHash(preMergeHeaders.getLast().getHash())
-                              .ifPresent(context::setIsPostMerge));
-                }
-              })
-          .thenCompose(ignored -> CompletableFuture.completedFuture(postMergeHeaders));
-    }
-  }
-
-  private CompletableFuture<List<BlockHeader>> storeBlockHeaders(
-      final List<BlockHeader> blockHeaders) {
-    return CompletableFuture.runAsync(
-            () -> {
-              final AtomicReference<Difficulty> difficulty =
-                  new AtomicReference<>(
-                      blockchain.calculateTotalDifficulty(blockHeaders.getFirst()));
-              blockHeaders.forEach(
-                  blockHeader -> {
-                    blockchain.unsafeStoreHeader(blockHeader, difficulty.get());
-                    difficulty.set(difficulty.get().add(blockHeader.getDifficulty()));
-                  });
-            })
-        .thenCompose(
-            ignored -> {
-              logProgress(blockHeaders.getLast());
-              return CompletableFuture.completedFuture(Collections.emptyList());
-            });
-  }
-
-  private void logProgress(final BlockHeader blockHeader) {
-    if (blockHeader.getNumber() >= lastPoWBlockNumber) {
-      LOG.info("Pre-merge headers import completed at block {}", lastPoWBlockNumber);
-    } else if (shouldLog.get()) {
-      throttledLog(
-          LOG::info,
-          String.format(
-              "Pre-merge headers import progress: %d of %d (%.2f%%)",
-              blockHeader.getNumber(),
-              lastPoWBlockNumber,
-              (double) (100 * blockHeader.getNumber()) / lastPoWBlockNumber),
-          shouldLog,
-          LOG_REPEAT_DELAY_SECONDS);
-    }
-  }
-}
